Add bookworm to initial build
The lint build changed to build off bookworm, but the way this was
structured it could (and was) running before bookworm is built and
updated, which meant it would fail because the lint job doesn't do an
apt-update (as it assumes it has just been built from an updated
bookworm base).

This fixes it by doing it (along with the existing initial debian/stable
base build) before any of the other builds.