Work around docker suck
Apparently when you run `docker build` it makes a full copy of `cwd`
into some tmp build path for no good reason, and then later on breaks if
anything in the original source directory has changed, which makes it
racy as hell (and also a huge waste of time if there is anything
substantial currently in the build directory).

This works around docker's garbage design by putting the generated
docker files into temp directories and running the build inside that
temp directory.